requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
streamlit>=1.28.0
requests>=2.31.0

torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
tqdm>=4.65.0
scikit-learn>=1.3.0
